usertest: extract ipc_util/ helpers (ns, fs, time)
Pull the raw-IPC helpers used across phases (ns_lookup/ns_stat/
ns_readdir, fs_create/mkdir/remove/rename/write_inline/read_bytes,
usertest_dir_cap, epoch_to_ymdhms) out of the omnibus main.rs into
ipc_util/{ns,fs,time}.rs. First step of the #21 structural refactor;
behaviour unchanged. main.rs shrinks ~270 lines.

Verified: cargo xtask build + run on both x86_64 and riscv64; usertest
emits ALL TESTS PASSED unchanged.
